requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
PyJWT[crypto]>=2.8.0,<3.0.0
pytest>=8.0.0,<9.0.0
httpx>=0.27.0,<1.0.0
